+++ HVAC_LIB/src/socket_reader.cpp
@@ -47,31 +47,31 @@ size_t SOCKET_READER::read( int _fd ) throw( exception )
 {
 	ssize_t rc = recv( _fd, this->read_buffer, GC_BUFFER_SIZE, MSG_DONTWAIT );
 
-	if( rc == -1 )
+	if ( rc == -1 )
 	{
-		throw( EXCEPTIONS::CONNECTION_ERROR( create_perror_string( "Failed to read from client:" ) ) );
+		throw ( EXCEPTIONS::CONNECTION_ERROR( create_perror_string( "Failed to read from client:" ) ) );
 	}
 
-	if( rc == 0 )
+	if ( rc == 0 )
 	{
 		/*
 		 * Client has disconnected
 		 */
-		throw( EXCEPTIONS::CONNECTION_ERROR( "Client connection closed." ) );
+		throw ( EXCEPTIONS::CONNECTION_ERROR( "Client connection closed." ) );
 	}
 
 	size_t start = 0;
-	size_t nl_idx = 0;
+	ssize_t nl_idx = 0;
 	size_t last_nl = 0;
 
-	while( ( nl_idx = this->find_nl_in_buffer( start, rc ) ) != ( size_t ) - 1 )
+	while ( ( nl_idx = this->find_nl_in_buffer( start, ( size_t )rc ) ) !=  - 1 )
 	{
-		this->line_vector.push_back( string( this->read_buffer, start, ( nl_idx - start ) + 1 ) );
-		last_nl = nl_idx;
-		start = nl_idx + 1;
+		this->line_vector.push_back( string( this->read_buffer, start, ( ( ( size_t )nl_idx ) - start ) + 1 ) );
+		last_nl = ( size_t )nl_idx;
+		start = ( size_t )nl_idx + 1;
 	}
 
-	if( last_nl < ( size_t )( rc - 1 ) )
+	if ( last_nl < ( size_t )( rc - 1 ) )
 	{
 		cerr << "Failed to find last NL in read buffer (" << last_nl << "|" << rc << ")" << endl;
 	}
@@ -85,9 +85,9 @@ size_t SOCKET_READER::get_line_count( void ) const
 }
 string SOCKET_READER::pop_first_line( void )
 {
-	if( this->get_line_count() < 1 )
+	if ( this->get_line_count() < 1 )
 	{
-		throw( runtime_error( "Line vector is empty." ) );
+		throw ( runtime_error( "Line vector is empty." ) );
 	}
 
 	string ret = this->line_vector[0];
@@ -95,13 +95,13 @@ string SOCKET_READER::pop_first_line( void )
 	return ret;
 }
 
-size_t SOCKET_READER::find_nl_in_buffer( size_t _start, size_t _end )
+ssize_t SOCKET_READER::find_nl_in_buffer( size_t _start, size_t _end )
 {
-	for( size_t i = _start; i < _end; i++ )
+	for ( size_t i = _start; i < _end; i++ )
 	{
-		if( read_buffer[i] == '\n' )
+		if ( read_buffer[i] == '\n' )
 		{
-			return i;
+			return ( ssize_t )i;
 		}
 	}
